Store order and cart to database

+++ app/src/main/java/test/s160419098/anmp/wu/data/Category.kt
@@ -7,7 +7,8 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "categories")
 data class Category(
     @PrimaryKey(autoGenerate = true)
-    val id: Int,
+    val id: Long,
+
     val name: String,
 ) {
     @Ignore
